requirements.txt
requests>=2.28.0,<2.32
python-dotenv>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
scipy>=1.7.0

# Opsiyonel: LLM ve model kütüphaneleri (Python 3.8 uyumlu)
openai>=1.0.0
anthropic>=0.70.0,<0.84
transformers>=4.0.0; python_version>="3.9"

